webapi/tct-netinfo-w3c-tests/netinfo/NetworkInformation_connection_readonly.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Fan,Weiwei <weiwix.fan@intel.com>
        Xinx, liu <xinx.liu@intel.com>

-->

<html>
  <head>
    <title>NetworkInfo Test: NetworkInformation_connection_readonly</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/WD-netinfo-api-20121129/#widl-NetworkInformation-connection" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if the NetworkInformation.connection attribute is readonly" />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
  <div id="log"></div>
    <script>
        test(function() {
            if (navigator.connection) {
                var init_1 = navigator.connection;
                navigator.connection = null;
                assert_true(navigator.connection == init_1 && navigator.connection != null, "navigator.connection is readonly");
            } else if(navigator.mozConnection) {
                var init_2 = navigator.mozConnection;
                navigator.mozConnection = null;
                assert_true(navigator.mozConnection == init_2 && navigator.mozConnection != null, "navigator.connection is readonly");
            } else if(navigator.webkitConnection) {
                var init_3 = navigator.webkitConnection;
                navigator.webkitConnection = null;
                assert_true(navigator.webkitConnection == init_3 && navigator.webkitConnection != null, "navigator.connection is readonly");
            } else {
                assert_true(false, "NetworkInformation.connection attribute exists");
            }
        }, document.title);
    </script>
  </body>
</html>
